+++ R/method-predict.R
@@ -366,13 +366,13 @@ predict.inlavaan_internal <- function(
         lavmodel_x <- lavaan::lav_model_set_parameters(lavmodel, xx)
         lavimplied <- lavaan::lav_model_implied(lavmodel_x)
 
-        LAMBDA <- lavaan___lav_model_lambda(lavmodel = lavmodel_x)
-        VETA <- lavaan___lav_model_veta(lavmodel = lavmodel_x)
-        EETA <- lavaan___lav_model_eeta(
+        LAMBDA <- lavaan:::lav_model_lambda(lavmodel = lavmodel_x)
+        VETA <- lavaan:::lav_model_veta(lavmodel = lavmodel_x)
+        EETA <- lavaan:::lav_model_eeta(
           lavmodel = lavmodel_x,
           lavsamplestats = lavsamplestats
         )
-        EY <- lavaan___lav_model_ey(
+        EY <- lavaan:::lav_model_ey(
           lavmodel = lavmodel_x,
           lavsamplestats = lavsamplestats
         )
@@ -399,8 +399,10 @@ predict.inlavaan_internal <- function(
           group.idx <- (g - 1) * nlevels + seq_len(nlevels)
           implied.group <- lapply(lavimplied, function(x) x[group.idx])
 
-          # positional: first argument is Lp in lavaan < 0.7 but lp in >= 0.7
-          decomp <- lavaan___lav_mvnorm_cluster_implied22l(Lp, implied.group)
+          decomp <- lavaan:::lav_mvn_cl_implied22l(
+            lp = Lp,
+            implied = implied.group
+          )
           MB.j <- compute_ml_ranef(y_g, Lp, decomp)
 
           ov.idx <- Lp$ov.idx
@@ -588,13 +590,13 @@ predict.inlavaan_internal <- function(
         lavmodel_x <- lavaan::lav_model_set_parameters(lavmodel, xx)
         lavimplied <- lavaan::lav_model_implied(lavmodel_x)
 
-        LAMBDA <- lavaan___lav_model_lambda(lavmodel = lavmodel_x)
-        VETA <- lavaan___lav_model_veta(lavmodel = lavmodel_x)
-        EETA <- lavaan___lav_model_eeta(
+        LAMBDA <- lavaan:::lav_model_lambda(lavmodel = lavmodel_x)
+        VETA <- lavaan:::lav_model_veta(lavmodel = lavmodel_x)
+        EETA <- lavaan:::lav_model_eeta(
           lavmodel = lavmodel_x,
           lavsamplestats = lavsamplestats
         )
-        EY <- lavaan___lav_model_ey(
+        EY <- lavaan:::lav_model_ey(
           lavmodel = lavmodel_x,
           lavsamplestats = lavsamplestats
         )
@@ -624,8 +626,10 @@ predict.inlavaan_internal <- function(
           group.idx <- (g - 1) * nlevels + seq_len(nlevels)
           implied.group <- lapply(lavimplied, function(x) x[group.idx])
 
-          # positional: first argument is Lp in lavaan < 0.7 but lp in >= 0.7
-          decomp <- lavaan___lav_mvnorm_cluster_implied22l(Lp, implied.group)
+          decomp <- lavaan:::lav_mvn_cl_implied22l(
+            lp = Lp,
+            implied = implied.group
+          )
           MB.j <- compute_ml_ranef(y_g, Lp, decomp)
 
           ov.idx <- Lp$ov.idx